pyproject.toml
[project]
name = "hi-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]
